job #11491 cleaned up MC-Java plugin; got rid of build warnings in all mc plugins
