+++ src/org.xtuml.bp.mc.c.source/plugin.xml
@@ -18,7 +18,7 @@
   </extension>
   <extension
         id="org.xtuml.bp.mc.mc3020.mc3020_builder"
-        name="MC-3020 (C Model Compiler)"
+        name="MC-3020 Builder"
         point="org.eclipse.core.resources.builders">
     <builder>
       <run class="org.xtuml.bp.mc.mc3020.Mc3020CSourceBuilder" />
