+++ src/org.xtuml.bp.mc.java.source/build.properties
@@ -1,13 +1,7 @@
 output.mcjava.jar = bin/
 bin.includes = mcjava.jar,\
-               build_settings/,\
-               defaults/,\
                META-INF/,\
                plugin.xml,\
                about.html
 source.mcjava.jar = src/
-src.includes = src/,\
-               plugin.xml,\
-               defaults/,\
-               build_settings/,\
-               META-INF/
+jre.compilation.profile = JavaSE-1.6
