+++ src/org.xtuml.bp.mc.java.source/plugin.xml
@@ -1,44 +1,35 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?eclipse version="3.4"?>
 <plugin>
-
-   <extension
-         id="MCNature"
-         name="model compiler nature"
-         point="org.eclipse.core.resources.natures">
-      <runtime>
-         <run
-               class="org.xtuml.bp.mc.java.source.MCNature">
-         </run>
-      </runtime>
-      <builder
-            id="org.xtuml.bp.mc.java.source.export_builder">
-      </builder>
-   </extension>
-   <extension
-         point="org.eclipse.ui.ide.projectNatureImages">
-      <image
-            icon="../org.xtuml.bp.core/icons/xtuml-nature.gif"
-            natureId="org.xtuml.bp.mc.java.source.MCNature"
-            id="org.xtuml.resoures.natureimage">
-      </image>
-   </extension>
-   <extension
-         id="export_builder"
-         name="Model Compiler Pre-Build"
-         point="org.eclipse.core.resources.builders">
-      <builder>
-         <run
-               class="org.xtuml.bp.mc.java.source.ExportBuilder">
-         </run>
-      </builder>
-   </extension>
-   <extension
-         point="org.xtuml.bp.core.model-compilers">
-      <model-compiler
-            name="Java Model Compiler"
-            class="org.xtuml.bp.mc.java.source.MCNewProjectWizard">
-      </model-compiler>
-   </extension>
-
+  <extension
+      id="mcjavanature"
+      name="model compiler nature"
+      point="org.eclipse.core.resources.natures">
+    <runtime>
+      <run class="org.xtuml.bp.mc.java.McJavaNature"></run>
+    </runtime>
+    <builder id="org.xtuml.bp.mc.java.mcjava_builder"></builder>
+  </extension>
+  <extension
+      point="org.eclipse.ui.ide.projectNatureImages">
+    <image
+      icon="../org.xtuml.bp.core/icons/xtuml-nature.gif"
+      natureId="org.xtuml.bp.mc.java.mcjavanature"
+      id="org.xtuml.resoures.natureimage">
+    </image>
+  </extension>
+  <extension
+      id="org.xtuml.bp.mc.java.mcjava_builder"
+      name="MC-Java Builder"
+      point="org.eclipse.core.resources.builders">
+    <builder>
+      <run class="org.xtuml.bp.mc.java.McJavaBuilder"></run>
+    </builder>
+  </extension>
+  <extension point="org.xtuml.bp.core.model-compilers">
+    <model-compiler
+      name="MC-Java"
+      class="org.xtuml.bp.mc.java.MCNewProjectWizard">
+    </model-compiler>
+  </extension>
 </plugin>
