+++ src/org.xtuml.bp.mc.masl/build.properties
@@ -2,9 +2,4 @@ source.masl.jar = src/
 output.masl.jar = bin/
 bin.includes = plugin.xml,\
                META-INF/,\
-               build_settings/,\
                masl.jar
-src.includes = src/,\
-               plugin.xml,\
-               build_settings/,\
-               META-INF/
