--- src/org.xtuml.bp.mc.masl/build_settings/build_setting.properties
@@ -1,9 +1,0 @@
-# These settings are used by the MC-3020 builder
-
-build_command =
-marking_file_location = defaults/marking
-launch_spec_location = defaults/launch_specification
-generated_code_dest = code_generation
-generated_source_code_dest = src
-additional_args_for_builder = -e
-xbuild_local_location = mc/bin/xtumlmc_build
